feat: migrate persistence from SQLite to PostgreSQL 17
- gorm.io/driver/postgres + pgx stdlib (pure Go, static build retained);
  dropped glebarez sqlite drivers
- config: DATABASE_URL (or DB_HOST/USER/PASSWORD/NAME); migrate runner now
  PostgreSQL dialect ($1, ON CONFLICT, boolean false); waitForDB retry on boot
- migrations: REAL→DOUBLE PRECISION; new 000003 library_plants + app_meta
- plant library is now the library_plants table (system of record), seeded
  idempotently from the JSON sources (version-hashed in app_meta); search/
  list/categories/count served by SQL with btree + pg_trgm indexes; hot
  per-plant lookups use an in-memory cache hydrated from the table.
  Diacritic-correct case-insensitive search via Go-lowercased search_text.
- deploy: dedicated postgres:17 sibling container (healthcheck, ./data/postgres
  volume), app waits on it; deploy.sh generates POSTGRES_PASSWORD in .env
- migrate test now runs against TEST_DATABASE_URL (skips without it)
- 13,889 plants live on Postgres; full smoke suite green on prod

+++ backend/migrations/000001_initial.up.sql
@@ -23,7 +23,7 @@ CREATE TABLE gardens (
     name          TEXT NOT NULL,
     description   TEXT NOT NULL DEFAULT '',
     location_type TEXT NOT NULL DEFAULT 'outdoor',
-    area_sqm      REAL,
+    area_sqm      DOUBLE PRECISION,
     created_at    TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at    TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     deleted_at    TIMESTAMP
@@ -57,7 +57,7 @@ CREATE TABLE care_log (
     user_id            TEXT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
     action             TEXT NOT NULL,
     note               TEXT NOT NULL DEFAULT '',
-    quantity_harvested REAL,
+    quantity_harvested DOUBLE PRECISION,
     timestamp          TIMESTAMP NOT NULL,
     created_at         TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
